Move clickableItem
